optional crc byte flip

+++ modbus_client.h
@@ -17,7 +17,7 @@ public:
 	std::vector<uint8_t> build_read_03(uint8_t addr, uint16_t reg,uint16_t cnt);
 	std::vector<uint8_t> build_write_reg_06(uint8_t addr, uint16_t reg,uint16_t val);
 	std::vector<uint8_t> build_write_multreg_16(uint8_t addr, uint16_t startreg,std::vector<uint16_t> vals);
-	uint16_t calc_crc(uint8_t *arr, uint8_t length);
+	uint16_t calc_crc(uint8_t *arr, int length);
 };
 //----------------------------------------------------------------------------------------------------------------------
 #endif /*MODBUS_CLIENT_H*/
